projects/P5_ChartMaker/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.0
Pillow==10.4.0

